Do not fail if type ID (e.g. default website type) is not set
Previously in `CRM_Xcm_Configuration` type IDs were casted to int,
resulting in `NULL` being cast to `0`. This is an invalid ID and results
in errors like "'0' is not a valid option for field website_type_id".
Now NULL is returned instead of `0`.

The default website type is now marked as required in the settings form.

The deprecated call of `CRM_Utils_Array::value()` is replaced and minor
style changes are performed.